first refactor of main frame
