+++ example/lib/HinarUI/include/modules/module_com.h
@@ -1,0 +1,6 @@
+#ifndef MODULE_COM_H
+#define MODULE_COM_H
+
+#define COM_BAUD_RATE       115200
+
+#endif // MODULE_COM_H
