+++ example/lib/HinarUI/include/modules/module_default.h
@@ -1,0 +1,7 @@
+#ifndef MODULE_DEFAULT_H
+#define MODULE_DEFAULT_H
+
+extern int module_UICORE_page;
+extern const int module_UICORE_totalPages;
+
+#endif // MODULE_DEFAULT_H
