+++ example/lib/HinarUI/include/modules/module_sht30.h
@@ -1,0 +1,11 @@
+#ifndef MODULE_SHT30_H
+#define MODULE_SHT30_H
+
+#include <Adafruit_SHT31.h>
+#include <Wire.h>
+
+extern Adafruit_SHT31 SHT;
+
+bool SHT30_Setup(void);
+
+#endif // MODULE_SHT30_H
