+++ example/platformio.ini
@@ -19,3 +19,6 @@ lib_deps =
     adafruit/Adafruit SSD1306 @ ^2.5.1
     adafruit/Adafruit GFX Library @ ^1.10.10
     adafruit/Adafruit SHT31 Library @ ^2.2.2
+
+build_flags =
+  -D BOARD_ESP_WROVER_KIT
