+++ example/src/main.cpp
@@ -1,5 +1,5 @@
 #include <Arduino.h>
-#include "UI.h"
+#include "HinarUI_Core.h"
 
 void setup() {
     menu.create();
